+++ src/routers/index.js
@@ -4,11 +4,11 @@ const router = express.Router();
 const userRouter = require('./userRouter');
 const productRouter = require('./productRouter');
 const reviewRouter = require('./reviewRouter');
-const cartsRouter = require('./cartsRouter');
+const cartRouter = require('./cartRouter');
 
 router.use('/users', userRouter);
 router.use('/list', productRouter);
 router.use('/review', reviewRouter);
-router.use('/carts', cartsRouter);
+router.use('/cart', cartRouter);
 
 module.exports = router;
